gene x candidate enhancers interactions upload
The gene x candidate enhancers interactions we used to call hits from the 'at-scale' experiment:

There are 78,562 interactions here, but we state we tested 78,776 in the paper - this latter sum total refers to unique sites targeted across both the paper's pilot and 'at-scale' experiments. In the "at scale" experiment from which we drew hits, there were 78,562 unique interactions. 

Additionally, we state we targeted 5,779 in candidate enhancers in the 'at-scale' experiment: in hindsight, this should've been stated as 5,723 (56 candidate enhancers we targeted did not have any genes above the cell % expression threshold in the surrounding 2Mb region. Please see manuscript for details re: expression threshold). Only 5,723 are included in this file.